30/09/23 - 2/10/23: Enemy and UI fixes
The enemy no longer breaks the game when you enter the Enemy testing zone (which also means that the dialogue zone should also be theoretically fixed), and that the User Interface connected to the player node (Pause menu specifically) is set directly to the screen, and not connected to the player. I have also "implemented" a Player Attack Indicator that should showcase what type of Attack the player shall use by setting up an animation player to the AttackTypeIndicator (also in the new User Interface Canvas Layer, connected to the game scene directly).